Add files via upload
This project uses a sound sensor as input, checks its reading against a threshold and if the value is greater than the threshold, it changes the servos target angle from maxAngle to minAngle or vice versa.   

The controller and the servos are coded separately but use the same input. The controller doesn’t have any hard limitation and will go up when the threshold is reached, then it will hold it for a bit then will start falling/ move down. That's all there is.  

With the servos, a major challenge was the power supply as the servo and Arduino shared power, with maximum output the servo could drain all power and send Arduino into a hard reboot. To solve this, a speed limit is coded, so the servos move slower and guarantees the power supply to Arduino. Another mechanism to prevent power drainage was to have a cool down period between each wing flap because continuous running of the servo also causes the power problem. With both mechanisms implemented the Arduino is guaranteed to have enough power to run and connect to Bluetooth